Fix TypeError with {{#cargo_compound_query:}} on PHP 8.0
PHP 8.0 now throws a TypeError if count() is called with something
that's not an array or Countable (in this case, null). This causes Cargo
compound queries to break as they utilize CargoQueryDisplayer in a way
that mFieldDescriptions is left uninitialized. Fix it by explicitly
specifying the empty array as the default value for field descriptions.

CargoQueryDisplayer proved challenging to test in isolation, so I
created a test class with basic integration tests that allow exercising
the full Cargo parser function stack, which allows testing the output
from CargoQueryDisplayer.

Bug: T327339
Change-Id: I252606d54cec213585d13916798547db2c20baf5