MAIN-28541 | Fix recreating Cargo tables with more than 500 usages
The `recreatedata` action uses a recursive JS function to recursively
schedule jobs to recreate data for a Cargo table via an API call. It
consults the count of incoming links to the template(s) declaring the
table to be recreated to determine how many calls to make, but this
count is incorrect on MW 1.39 because the code doesn't work well with
the newly normalized templatelinks table. Update the code to fix the
query.