PLATFORM-9063 | Only run CargoStore validation prior to attempting to store data
CargoStore currently runs validation logic on incoming data each time
it's invoked, triggering a primary DB connection to fetch information
about the table and currently stored page data. This is only useful in
cases like a page save or data recreation, as those are the only cases
where incoming data will be stored - this logic should not be run on
parser cache misses for a regular page view, as primary DB connections
should be avoided there. So, only perform this validation when the data
would actually be stored.

Bug: T357395
Change-Id: Ia4e783a5707a08486d73fad6a8a374bdc646c86d