PLATFORM-9062 | Replace use of PRIMARY_DB connection in CargoUtils, CargoAttach, CargoDeclare and CargoPageValues
Some of the actions are still using PRIMARY DB connection during the Read requests, and therefore we should replace them with use of the proper REPLICA DB. All of those cases are simple and shouldn't yield any errors when PRIMARY DB is not used to query the results.
Also, mark CargoRecreateTablesAPI as writing API to minimise noise from transaction profile.