PLATFORM-9121 | Properly handle setting character set for connections
Now that we may use MW's LB abstractions for managing Cargo DB
connections on some cases, we must account for the fact that these tend
to return DBConnRef instances instead of raw connection handles. Unwrap
DBConnRefs accordingly so that we can operate on the backing connection
directly when setting the charset.